finished ch 7

+++ ray.hpp
@@ -4,6 +4,7 @@
 #include "vector3.hpp"
 
 
+// Every ray gets a starting point (origin) and a direction ("endpoint")
 class ray {
     public:
         ray() {}
@@ -11,7 +12,7 @@ class ray {
         
         point3 get_origin() const { return origin; }
         vec3 get_direction() const { return direction; }
-        point3 point_at_parameter(float t) const { return operator+(origin, operator*(direction, t)); }
+        point3 point_at_parameter(double t) const { return operator+(origin, operator*(direction, t)); }
 
         point3 origin;
         vec3 direction;
