+++ vector3.hpp
@@ -85,7 +85,7 @@ inline vec3& vec3::operator*=(const double t) {
 }
 
 inline vec3& vec3::operator/=(const double t) {
-    float k = 1.0 / t;
+    double k = 1.0 / t;
 
     e[0] *= k;
     e[1] *= k;
@@ -94,7 +94,7 @@ inline vec3& vec3::operator/=(const double t) {
 }
 
 inline void vec3::make_unit_vector() {
-    float k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
+    double k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
     
     e[0] *= k;
     e[1] *= k;
@@ -119,15 +119,15 @@ inline vec3 operator/(const vec3 &v1, const vec3&v2) {
     return vec3(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1], v1.e[2] / v2.e[2]);
 }
 
-inline vec3 operator*(const vec3 &v, float t) {
+inline vec3 operator*(const vec3 &v, double t) {
     return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-inline vec3 operator/(const vec3 &v, float t) {
+inline vec3 operator/(const vec3 &v, double t) {
     return vec3(v.e[0] / t, v.e[1] / t, v.e[2] / t);
 }
 
-inline float dot(const vec3 &v1, const vec3 &v2) {
+inline double dot(const vec3 &v1, const vec3 &v2) {
     return (v1.e[0] * v2.e[0]) + (v1.e[1] * v2.e[1]) + (v1.e[2] * v2.e[2]);
 }
